Fix buck2 compilation errors by moving rebaser-client tests
Fix buck2 compilation errors by moving rebaser-client tests from unit
tests in the rebaser-core crate to integration tests in the
rebaser-server crate.

Not only does this make more sense from a domain perspective (why would
rebaser-core host thoses tests?), but now we can prevent cyclic
dependencies by using the integration test buck2 rule for clean
separation.
