+++ lib/rebaser-server/Cargo.toml
@@ -32,3 +32,4 @@ veritech-client = { path = "../../lib/veritech-client" }
 [dev-dependencies]
 dal-test = { path = "../../lib/dal-test" }
 pretty_assertions_sorted = { workspace = true }
+rebaser-client = { path = "../../lib/rebaser-client" }
