+++ lib/rebaser-server/tests/integration_test/mod.rs
@@ -1,1 +1,2 @@
+mod client;
 mod connection;
